+++ libs/ros1bridge/src/point_cloud2.cpp
@@ -167,8 +167,7 @@ bool mrpt::ros1bridge::fromROS(const sensor_msgs::PointCloud2& msg, CPointsMapXY
       get_float_from_field(z_field, msg_data, z);
       get_float_from_field(i_field, msg_data, i);
       obj.insertPoint(x, y, z);
-
-      obj.setPointIntensity(obj.size() - 1, i);
+      obj.insertPointField_Intensity(i);
     }
   }
   return true;